Move components to a correct folder
