+++ components/PlayerPage/PlayerDetails.tsx
